% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Retro_data}
\alias{Retro_data}
\title{Simulated joint modeling data}
\format{
A data frame with 200 rows and 33 variables:
}
\usage{
Retro_data
}
\description{
A dataset contains 200 sample. The dataset is in wide format, with 13 longitudinal measurement
and corresponding time points.
}
\keyword{datasets}
