% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetroJM.R
\name{RetroJM}
\alias{RetroJM}
\title{Joint model on the retrospective time scale}
\usage{
RetroJM(
  data,
  long_format = TRUE,
  outcome_var,
  time_var,
  Ai_var,
  last_obs_time_var,
  cov_var,
  ID_var,
  surv_censor_var,
  drop_censor_var,
  longi_end_time = NULL,
  surv_drop_knots = 0,
  longi_inter_knots = 0,
  num_cores = detectCores() - 2
)
}
\arguments{
\item{data}{The dataset used in analysis. It could be either long format or wide format. 
It should contain longitudinal measurement, measurement time, a binary variable as 
parameter of interest, ID for subjects, event/censoring time, an indicator
for death status, an indicator for drop out status, and any covariates.}

\item{long_format}{A logical argument takes value TRUE or FALSE. If TRUE, the input dataset
is in long format. If FALSE, the input dataset should be wide format. Defaut is TRUE.}

\item{outcome_var}{Name of outcome variables. If the input data is long format, it should 
be a single character; If the input is wide format, it should be a character vector.}

\item{time_var}{Name of observed time variables. If the input data is long format, it should 
be a single character; If the input is wide format, it should be a character vector. It should 
have same length as "outcome_var".}

\item{Ai_var}{Name of parameter of interest. Have to be a binary variable coded as 0/1, 
such as control/treatment.}

\item{last_obs_time_var}{Name of variable for event/censoring time.}

\item{cov_var}{Name of covariates. Default is NULL, meaning no covariates.}

\item{ID_var}{Name of ID variable.}

\item{surv_censor_var}{Name of indicator variable for death status.}

\item{drop_censor_var}{Name of indicator variable for drop out status.}

\item{longi_end_time}{The end time for longitudinal trajectory. It take numeric as input.
If NULL, then it equals to study end time. Default is NULL.}

\item{surv_drop_knots}{The number of additional pieces for piecewise constant hazard model in survival
and dropout hazard. Default is 0, meaning no additional pieces and the hazard is constant.}

\item{longi_inter_knots}{The number of additional pieces for piecewise linear model in 
longitudinal trajectory. Default is 0, meaning no additional pieces and the longitudinal
trajectory is linear.}

\item{num_cores}{The number of cores used for parallel computing.}
}
\value{
A `list` of `3` datasets containing the parameter estiamtes for survival, dropout, and 
longitudinal measurement.
Each dataset has row representing each parameter, 5 columns for `Estimates`, `Standard Error`,
`Lower bound for 95\% Confidence Interval`, `Upper bound for 95\% Confidence Interval`,
`P value`.
}
\description{
\loadmathjax
RetroJM is used for fitting joint model on retrospective time scale starting from death time.
The outcome model is:
\mjdeqn{Y_i(t^*)=\beta_{\mu}(t^*)+A_i\beta_A(t^*)+X_i^T\psi_X+U_i^*(t^*)+\epsilon_i(t^*)}{}
where \mjeqn{t^*=D-t}{}, D is death time and t is observed time. 
The hazard for competing risks are: 
\mjdeqn{\lambda_i^{\delta}(t)=\exp{\Big(\alpha_0^{\delta}(t)+A_i\alpha_A^{\delta}(t)+\alpha_U^{\delta} U_i \Big)},\hspace{0.2cm}\text{for}\hspace{0.12cm} \delta=1,2,}{}
}
\examples{

\dontrun{
library(RetroJM)

## load the simulated example data Retro_data
data(Retro_data)

test_run<-RetroJM(data = Retro_data,
long_format = FALSE,
outcome_var = paste0("Y",1:13),
time_var = paste0("time",1:13),
Ai_var = "Ai",
last_obs_time_var = "observed.t",
cov_var = paste0("x",1:2),
ID_var = "ID",
surv_censor_var = "surv_ind",
drop_censor_var = "drop_ind",
longi_end_time = 25,
surv_drop_knots = 1,
longi_inter_knots = 1,
num_cores = 2)

}


}
